Genero "Acabado"
Por parte de la api tabla genero acabada

+++ api-anineo/src/generos/generos.service.ts
@@ -1,9 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable, HttpStatus } from '@nestjs/common';
 import { Genero } from './genero.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateGeneroDto } from './dto/create_genero.dto';
-
 
 @Injectable()
 export class GenerosService {
@@ -11,20 +10,39 @@ export class GenerosService {
         @InjectRepository(Genero) private generoRepository: Repository<Genero>
     ){}
 
-    createGenero(genero: CreateGeneroDto){
-        return this.generoRepository.save(this.generoRepository.create(genero));
+    async createGenero(genero: CreateGeneroDto){
+        const generoExist= await this.generoRepository.findOne({
+            where: {
+                nombre: genero.nombre
+            }
+        })
+        if(generoExist){
+            return new HttpException('El usuario ya existe', HttpStatus.CONFLICT);
+        }else{
+            return this.generoRepository.save(this.generoRepository.create(genero));
+        }
     }
 
     findAll(): Promise<Genero[]> {
         return this.generoRepository.find();
     }
 
-    findOne(id: number): Promise<Genero | null> {
-        return this.generoRepository.findOneBy({ id });
+    async findOne(id: number){
+        const generoExist= await this.generoRepository.findOneBy({ id });
+        if(generoExist){
+            return generoExist;
+        }else{
+            return new HttpException('Genero inexistente', HttpStatus.NOT_FOUND);
+        }
     }
     
-    async remove(id: number): Promise<void> {
-        await this.generoRepository.delete(id);
+    async remove(id: number){
+        const generoExist= await this.generoRepository.findOneBy({ id });
+        if(!generoExist){
+            return new HttpException('Genero Inexistente', HttpStatus.NOT_FOUND);
+        }else{
+            return this.generoRepository.delete(id);
+        }
     }
 
     async updateGenero(id: number, createGenero: CreateGeneroDto): Promise<Genero> {
